requirements.txt
mutagen==1.46.0
python3-discogs-client==2.6
pick==2.2.0
